Update: Organisms のコンポーネント以下のファイル名をコンポーネント名に変更
