+++ src/components/organisms/MotionSensorFloatingButton/MotionSensorFloatingButton.tsx
@@ -1,4 +1,4 @@
-import React, { Dispatch, SetStateAction, useCallback } from "react";
+import React, { Dispatch, SetStateAction } from "react";
 
 // FontAwesome
 import SensorsIcon from "@mui/icons-material/Sensors";
@@ -6,6 +6,9 @@ import SensorsOffIcon from "@mui/icons-material/SensorsOff";
 
 // components
 import { FloatingButton } from "../../atoms/FloatingButton";
+
+// Hooks
+import { useMotionSensorFloatingButton } from "./MotionSensorFloatingButton.hooks";
 
 type Props = {
   isSensorUse: boolean;
@@ -16,13 +19,8 @@ export const MotionSensorFloatingButton: React.VFC<Props> = ({
   isSensorUse,
   setIsSensorUse,
 }) => {
-  const setIsClickedFalseOnClick = useCallback(() => {
-    setIsSensorUse(false);
-  }, [setIsSensorUse]);
-
-  const setIsClickedTrueOnClick = useCallback(() => {
-    setIsSensorUse(true);
-  }, [setIsSensorUse]);
+  const { setIsClickedFalseOnClick, setIsClickedTrueOnClick } =
+    useMotionSensorFloatingButton(setIsSensorUse);
 
   return isSensorUse ? (
     <FloatingButton icon={SensorsOffIcon} onClick={setIsClickedFalseOnClick} />
