+++ src/components/organisms/MotionSensorFloatingButton/index.ts
@@ -1,0 +1,2 @@
+export { MotionSensorFloatingButton } from "./MotionSensorFloatingButton";
+export { useMotionSensorFloatingButton } from "./MotionSensorFloatingButton.hooks";
